Modify the equals() method in the Book class and subclasses and update the TestEqualsAfterInheritance class
The choice for equals() methods is to have two different methods: a standard o1.equals(o2) method that returns true only if o1 and o2 are members of the same class and all the equalities holds. A class method CLASS.equals(o1, o2) that simply checks whether o1 and o2 are instances of CLASS and then checks for equalities. In this way it is possible to query if two objects are equals using different criteria, while maintaining the reflectivity property in the usual o1.equals(o2) method.